Add cli test usage Add Python cli convenient usage example and test
