+++ test/cli/generate_datakey_withoutplaint.py
@@ -38,6 +38,7 @@ def generate_datakey_withoutplaint(base_url, keyid, len, aad):
         return
 
     print('generate_datakey_withoutplaint resp:\n%s\n' %(resp.text))
+    return json.loads(resp.text)['result']['ciphertext']
 
 if __name__ == "__main__":
     headers = _utils_.headers
